Fix the signed tests errors, add the CPR-234 report

+++ solidity/simple/loop/range_bitlength_inference/u16_to_i16.sol
@@ -19,7 +19,7 @@ pragma solidity ^0.8.0;
 contract Test {
     function main() public pure returns(int16) {
         int16 sum = 0;
-        for(int16 i = -uint16(200); i < 200; i++) {
+        for(int16 i = -200; i < 200; i++) {
             sum += i;
         }
         return sum;
